app/requirements.txt
asgiref==3.4.1
Django==3.1.12
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
